Update at point where GodivaR matches.  Still a problem with PWR_16x16.

+++ src/RayTrace/gpuGlobal.h
@@ -12,7 +12,9 @@
 
 namespace MonteRay{
 typedef float gpuFloatType_t;
-typedef float gpuTallyType_t;
+
+typedef double gpuTallyType_t;
+//typedef float gpuTallyType_t;
 
 const gpuFloatType_t gpu_neutron_molar_mass = 1.00866491597f;
 const gpuFloatType_t gpu_AvogadroBarn = .602214179f;
